requirements.txt
numpy>=1.21.0
pandas>=1.3.0
opencv-python>=4.5.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
mlcroissant>=1.0.0
joblib>=1.1.0
streamlit>=1.28.0
pillow>=9.0.0
